build: fix lpthread dependency with glslang
"-pthread" is enough for a dynamic build. Indeed, the pthread API is used
within libplacebo, so the compiler know that it shoud link with pthread in that
case.

However, this may fail with static builds. If the code using libplacebo.a and
all glslang .a doesn't use the pthread API, the compiler won't link with
pthread. To work-around this issue, make libplacebo depends on -lpthread, if
present.